[FIX] request attendance manual

+++ database/migrations/2024_02_05_132514_create_timeoff_period_regulations_table.php
@@ -14,8 +14,8 @@ return new class extends Migration
         Schema::create('timeoff_period_regulations', function (Blueprint $table) {
             $table->id();
             $table->foreignId('timeoff_regulation_id')->constrained();
-            $table->unsignedTinyInteger('min_working_month')->default(0);
-            $table->unsignedTinyInteger('max_working_month');
+            $table->unsignedSmallInteger('min_working_month')->default(0);
+            $table->unsignedSmallInteger('max_working_month');
             $table->timestamps();
         });
     }
